+++ storage_image/models/__init__.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from . import storage_image
-from . import storage_thumbnail
 from . import partner
 from . import product
+from . import storage_image
 from . import storage_image_wizard
+from . import storage_image_factory
+from . import storage_thumbnail
 from . import storage_thumbnail_wizard
+from . import storage_thumbnail_factory
